Upgrade CurrentWeatherEntry int fields to doubles to prevent type errors

+++ app/src/main/java/com/tjclawson/mvvm_practice/data/response/CurrentWeatherEntry.kt
@@ -4,29 +4,29 @@ package com.tjclawson.mvvm_practice.data.response
 import com.google.gson.annotations.SerializedName
 
 data class CurrentWeatherEntry(
-    val cloudcover: Int,
-    val feelslike: Int,
-    val humidity: Int,
+    val cloudcover: Double,
+    val feelslike: Double,
+    val humidity: Double,
     @SerializedName("is_day")
     val isDay: String,
     @SerializedName("observation_time")
     val observationTime: String,
-    val precip: Int,
-    val pressure: Int,
-    val temperature: Int,
+    val precip: Double,
+    val pressure: Double,
+    val temperature: Double,
     @SerializedName("uv_index")
-    val uvIndex: Int,
-    val visibility: Int,
+    val uvIndex: Double,
+    val visibility: Double,
     @SerializedName("weather_code")
-    val weatherCode: Int,
+    val weatherCode: Double,
     @SerializedName("weather_descriptions")
     val weatherDescriptions: List<String>,
     @SerializedName("weather_icons")
     val weatherIcons: List<String>,
     @SerializedName("wind_degree")
-    val windDegree: Int,
+    val windDegree: Double,
     @SerializedName("wind_dir")
     val windDir: String,
     @SerializedName("wind_speed")
-    val windSpeed: Int
+    val windSpeed: Double
 )
